simplying query for indicators
