+++ .env.sample
@@ -1,1 +1,2 @@
 SASS_PATH=./node_modules/;./src
+REACT_APP_DATA_TABLENAME=covid_data_dev
