--- src/components/widget/widget-templates/bars/index.js
@@ -1,1 +1,0 @@
-export { default } from './component';
