--- src/components/widget/widget-templates/bars/styles.scss
@@ -1,17 +1,0 @@
-
-@font-face {
-  font-family: 'Chalet';
-  font-weight: $font-weight-medium;
-  src: font-url('/typography/ChaletNewYorkNineteenSixty.ttf') format('truetype');
-}
-
-.c-number-of-services {
-  p {
-    color: $color-5;
-    font-size: 13px;
-    margin: 20px 0;
-    font-weight: 600;
-  }
-}
-
-
