+++ src/services/categories.js
@@ -5,7 +5,9 @@ export const fetchCategories = () => {
     SELECT
       name,
       slug
-    FROM covid_categories`;
+    FROM covid_categories
+    ORDER by name`;
+
   return cartoApi(query);
 };
